Working remote

+++ libraries/common.rb
@@ -1,9 +1,9 @@
 def packageFile(mode)
   case mode
-  when :local
-    node.default.mesos.install.pkg_local
-  when :remote
-    "#{node.default.mesos.source.dir}/#{node.default.mesos.install.filename}"
+  when 'local'
+    node[:mesos][:install][:pkg_local]
+  when 'remote'
+    "#{node[:mesos][:source][:dir]}/#{node[:mesos][:install][:filename]}"
   end
 end
 
@@ -13,46 +13,63 @@ end
 
 
 # mode:{local|master|slave|other}
-def stopMesos(mode=:local)
+def serviceMesos(serviceAction='stop', mode='local')
+  service 'zookeeper' do
+    provider Chef::Provider::Service::Upstart
+    supports :status => true, :start => true, :stop => true, :restart => true
+    action serviceAction
+  end
+
   case mode
-  when :local
+  when 'local'
+    # Some distros around have this, we only support stop if so
+    # and overwrite it.
     service "mesos-local" do # Nope, kill it for good
-      supports :status => true
-      action :stop
+      provider Chef::Provider::Service::Upstart
+      supports :stop => true
+      only_if do serviceAction == 'stop' end 
+      action serviceAction
     end
-  when :master
+
+    serviceMesos serviceAction, 'master'
+    serviceMesos serviceAction, 'slave'
+
+  when 'master'
     service "mesos-master" do # All us
-      supports :status => true
-      action :stop
+      provider Chef::Provider::Service::Upstart
+      supports :status => true, :start => true, :stop => true, :restart => true
+      action serviceAction
     end
-  when :slave
+  when 'slave'
     service "mesos-slave" do # All us
-      supports :status => true
-      action :stop
+      provider Chef::Provider::Service::Upstart
+      supports :status => true, :start => true, :stop => true, :restart => true
+      action serviceAction
     end    
-  when :other
+  when 'other'
     # For safety, in-case someone installed chef over an existing install
     # by another script/app
     service "mesos" do # Not us, kill it for good
-      supports :status => true
-      action :stop
+      provider Chef::Provider::Service::Upstart
+      supports :status => true, :start => true, :stop => true, :restart => true
+      action serviceAction
     end
   end
 end
 
 # mode:{local|master|slave|other}
-def checkFile(mode=:local)
+def checkFile(mode='local')
   prefix = node[:mesos][:install][:prefix]
   local_prefix = node[:mesos][:install][:local_prefix]
 
   log "Mesos prefix is: [#{prefix}], local prefix is: [#{local_prefix}]"
 
   case mode
-  when :local
+  when 'local'
     "#{local_prefix}/#{node[:mesos][:install][:local_script]}"
-  when :master
+  when 'master'
     "#{prefix}/#{node[:mesos][:install][:master_script]}"
-  when :slave
+  when 'slave'
     "#{prefix}/#{node[:mesos][:install][:slave_script]}"
   else
     raise "Must supply 'mode' to check_file method"
@@ -60,150 +77,9 @@ def checkFile(mode=:local)
 end
 
 # mode:{local|master|slave|other}
-def mesosExists?(mode=:local)
+def mesosExists?(mode='local')
   mesos_file = checkFile mode
   log "using file: [#{mesos_file}]"
   File.exists?(mesos_file)
 end
 
-def installPkg
-  log "Installing via remote package..."
-
-  remote_file 'download_file' do
-    path packageFile :remote
-    source node.default.mesos.install.pkg_url
-    # owner 'vagrant'
-    # group 'vagrant'
-    # mode 00644
-    action :create_if_missing
-    notifies :install, "dpkg_package[mesos_deb]"
-  end
-end
-
-def installDefaultTemplate
-  template "/etc/default/mesos" do
-    source "mesos.erb"
-    mode 0440
-    owner "root"
-    group "root"
-    variables({
-       :log_location => node[:mesos][:conf][:log_location],
-       :ulimit => node[:mesos][:conf][:ulimit],
-       :zookeepers => node[:mesos][:conf][:zookeepers],
-       :masters => node[:mesos][:conf][:masters],
-       :options => node[:mesos][:conf][:options]
-    })
-  end
-end
-
-def installMasterConf
-  template "/etc/init/mesos-master.conf" do
-    source "mesos.erb"
-    mode 0440
-    owner "root"
-    group "root"
-    variables({
-       :log_location => node[:mesos][:conf][:log_location],
-       :ulimit => node[:mesos][:conf][:ulimit],
-       :zookeepers => node[:mesos][:conf][:zookeepers],
-       :masters => node[:mesos][:conf][:masters],
-       :options => node[:mesos][:conf][:options]
-    })
-  end
-end
-
-def installSlaveConf
-  template "/etc/init/mesos-slave.conf" do
-    source "mesos.erb"
-    mode 0440
-    owner "root"
-    group "root"
-    variables({
-       :log_location => node[:mesos][:conf][:log_location],
-       :ulimit => node[:mesos][:conf][:ulimit],
-       :zookeepers => node[:mesos][:conf][:zookeepers],
-       :masters => node[:mesos][:conf][:masters],
-       :options => node[:mesos][:conf][:options]
-    })
-  end
-end
-
-
-# def installTemplates(mode)
-#   case mode
-#   when :master
-#     template "/etc/sudoers" do
-#       source "sudoers.erb"
-#       mode 0440
-#       owner "root"
-#       group "root"
-#       variables({
-#          :sudoers_groups => node[:authorization][:sudo][:groups],
-#          :sudoers_users => node[:authorization][:sudo][:users]
-#       })
-#     end
-#   when :slave
-# end
-
-def installMesos
-  mode = node.mesos.install.via
-
-  case mode
-  when "pkg"
-    installPkg
-  when "pkg_local"
-    log "Installing via local package..."
-
-    if File.exists? packageFile :local
-      log "Local package file exists, using it for install..."
-
-      dpkg_package "mesos_deb_local" do
-        action :install
-        only_if do File.exists? packageFile :local end
-      end
-    else
-      # log "Resolve set to: #{node.mesos.install.resolve}"
-
-      if node.mesos.install.resolve
-        log "Resolve set, will try to install with remote package..."
-        installPkg
-      else
-        fatal "Could not find package file, resolve not set, nothing to do. Raising exception."
-      end
-    end
-
-  when "src"
-    log "Installing via Package"
-
-    git node.mesos.source.dir do
-      repository node.mesos.source.repo
-      reference node.mesos.source.branch
-      action :sync
-      notifies :run, "bash[configure_mesos]"
-    end  
-  end
-end
-
-def beginInstall(mode)
-  stopMesos mode
-  # TODO: move this into local.rb
-  if mesosExists? mode
-    if node.mesos.install.force
-      bannerLog "Force installing mesos."
-      installMesos
-      installDefaultTemplate
-      installMasterConf
-      installSlaveConf
-    else
-      log "Seems mesos-#{mode} already exists, not forcing - nothing to do."
-    end
-  else
-    converge_by("Create mesos-#{mode}...") do
-      bannerLog "Installing mesos."
-      installMesos
-      installDefaultTemplate
-      installMasterConf
-      installSlaveConf
-    end
-  end
-end
